feat: new server architecture
