+++ server/discords.json
@@ -1,0 +1,1 @@
+{ "0x1c0f0af3262a7213e59be7f1440282279d788335": "cha0s.g0d" }
